+++ utils/constants.js
@@ -1,3 +1,3 @@
-export const MAX_RULES_LIMIT = 5;
+export const MAX_RULES_LIMIT = 10;
 export const VERIFY_API_URL = 'https://blockdistraction.com/api/verifyKey';
 export const IS_FIREFOX = true;